pyproject.toml
[project]
name = "llm-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain>=0.3.26",
    "langchain-openai>=0.3.28",
    "python-dotenv>=1.1.1",
]
